tasks/terraform.yml: Cleanup the terraform before the build
This is because terraform package uses make fmt to fix the
formatting issue before actually call go build.  This is just
weired and shouldn't have any code which is not following the
gofmt.  Until we fix the code in the repo, we explicitly
update the local code with the remote, to avoid the go get
error, due to the local code change.